implemented persistence layer for alerts, packages renaming for better structure
