--- alert-processor-service/src/main/java/com/alertprocessorservice/AlertProcessorApplication.java
@@ -1,13 +1,0 @@
-package com.alertprocessorservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-class AlertProcessorApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AlertProcessorApplication.class, args);
-    }
-
-}
