--- alert-processor-service/src/main/java/com/alertprocessorservice/persistence/repository/TradeRepository.java
@@ -1,9 +1,0 @@
-package com.alertprocessorservice.persistence.repository;
-
-import com.alertprocessorservice.persistence.model.Trade;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TradeRepository extends JpaRepository<Trade, Long> {
-}
