+++ alert-processor-service/src/main/java/com/findataagg/alertprocessorservice/alert/model/ConditionType.java
@@ -1,0 +1,9 @@
+package com.findataagg.alertprocessorservice.alert.model;
+
+/**
+ * Defines the types of conditions for an alert rule.
+ */
+public enum ConditionType {
+    GREATER_THAN,
+    LESS_THAN,
+}
