+++ alert-processor-service/src/main/java/com/findataagg/alertprocessorservice/persistence/repository/TradeRepository.java
@@ -1,0 +1,9 @@
+package com.findataagg.alertprocessorservice.persistence.repository;
+
+import com.findataagg.alertprocessorservice.persistence.model.Trade;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TradeRepository extends JpaRepository<Trade, Long> {
+}
