+++ common/src/main/java/com/findataagg/common/messaging/model/PriceUpdate.java
@@ -17,6 +17,6 @@ public record PriceUpdate(@JsonProperty("S") String symbol,
     
     @Override
     public String type() {
-        return "price";
+        return "trade";
     }
 }
