--- data-ingestor-service/src/main/java/com/dataingestorservice/DataIngestorApplication.java
@@ -1,19 +1,0 @@
-package com.dataingestorservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-
-@SpringBootApplication
-@EnableConfigurationProperties
-public class DataIngestorApplication {
-
-    public static void main(String[] args) throws InterruptedException {
-        SpringApplication.run(DataIngestorApplication.class, args);
-
-        // This line will block the main thread, keeping the application alive
-        // so the WebSocket's background threads can continue listening for messages.
-        Thread.currentThread().join();
-    }
-
-}
