--- data-ingestor-service/src/main/java/com/dataingestorservice/websocket/AlpacaAuthRequest.java
@@ -1,4 +1,0 @@
-package com.dataingestorservice.websocket;
-
-public record AlpacaAuthRequest(String action, String key, String secret) {
-}
