--- data-ingestor-service/src/main/java/com/dataingestorservice/websocket/SubscriptionRequest.java
@@ -1,6 +1,0 @@
-package com.dataingestorservice.websocket;
-
-import java.util.List;
-
-public record SubscriptionRequest(String action, List<String> trades, List<String> quotes) {
-}
