+++ data-ingestor-service/src/main/java/com/findataagg/dataingestorservice/websocket/AlpacaAuthRequest.java
@@ -1,0 +1,4 @@
+package com.findataagg.dataingestorservice.websocket;
+
+public record AlpacaAuthRequest(String action, String key, String secret) {
+}
