+++ data-ingestor-service/src/main/java/com/findataagg/dataingestorservice/websocket/SubscriptionRequest.java
@@ -1,0 +1,6 @@
+package com.findataagg.dataingestorservice.websocket;
+
+import java.util.List;
+
+public record SubscriptionRequest(String action, List<String> trades, List<String> quotes) {
+}
