--- data-ingestor-service/src/test/java/com/dataingestorservice/DataIngestorApplicationTests.java
@@ -1,8 +1,0 @@
-package com.dataingestorservice;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DataIngestorApplicationTests {
-
-}
