+++ data-ingestor-service/src/test/java/com/findataagg/dataingestorservice/DataIngestorApplicationTests.java
@@ -1,0 +1,8 @@
+package com.findataagg.dataingestorservice;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DataIngestorApplicationTests {
+
+}
